refactor reduce
