+++ sparse/_coo/core.py
@@ -6,13 +6,14 @@ from functools import reduce
 import warnings
 
 import numpy as np
+import numba
 import scipy.sparse
 from numpy.lib.mixins import NDArrayOperatorsMixin
 
 from .._common import dot, matmul
 from .indexing import getitem
 from .umath import elemwise, broadcast_to
-from .._sparse_array import SparseArray
+from .._sparse_array import SparseArray, _reduce_super_ufunc
 from .._utils import normalize_axis, equivalent, check_zero_fill_value, _zero_of_dtype
 
 
@@ -658,6 +659,132 @@ class COO(SparseArray, NDArrayOperatorsMixin):  # lgtm [py/missing-equals]
         )
 
     __repr__ = __str__
+
+    def reduce(self, method, axis=(0,), keepdims=False, **kwargs):
+        """
+        Performs a reduction operation on this array.
+
+        Parameters
+        ----------
+        method : numpy.ufunc
+            The method to use for performing the reduction.
+        axis : Union[int, Iterable[int]], optional
+            The axes along which to perform the reduction. Uses all axes by default.
+        keepdims : bool, optional
+            Whether or not to keep the dimensions of the original array.
+        kwargs : dict
+            Any extra arguments to pass to the reduction operation.
+
+        Returns
+        -------
+        COO
+            The result of the reduction operation.
+
+        Raises
+        ------
+        ValueError
+            If reducing an all-zero axis would produce a nonzero result.
+
+        See Also
+        --------
+        numpy.ufunc.reduce : A similar Numpy method.
+        GCXS.reduce : Equivalent operation on GCXS arrays.
+
+        Examples
+        --------
+        You can use the :obj:`COO.reduce` method to apply a reduction operation to
+        any Numpy :code:`ufunc`.
+
+        >>> from sparse import COO
+        >>> x = np.ones((5, 5), dtype=np.int)
+        >>> s = COO.from_numpy(x)
+        >>> s2 = s.reduce(np.add, axis=1)
+        >>> s2.todense()  # doctest: +NORMALIZE_WHITESPACE
+        array([5, 5, 5, 5, 5])
+
+        You can also use the :code:`keepdims` argument to keep the dimensions after the
+        reduction.
+
+        >>> s3 = s.reduce(np.add, axis=1, keepdims=True)
+        >>> s3.shape
+        (5, 1)
+
+        You can also pass in any keyword argument that :obj:`numpy.ufunc.reduce` supports.
+        For example, :code:`dtype`. Note that :code:`out` isn't supported.
+
+        >>> s4 = s.reduce(np.add, axis=1, dtype=np.float16)
+        >>> s4.dtype
+        dtype('float16')
+
+        By default, this reduces the array by only the first axis.
+
+        >>> s.reduce(np.add)
+        <COO: shape=(5,), dtype=int64, nnz=5, fill_value=0>
+        """
+        axis = normalize_axis(axis, self.ndim)
+        zero_reduce_result = method.reduce([self.fill_value, self.fill_value], **kwargs)
+        reduce_super_ufunc = None
+
+        if not equivalent(zero_reduce_result, self.fill_value):
+            reduce_super_ufunc = _reduce_super_ufunc.get(method, None)
+
+            if reduce_super_ufunc is None:
+                raise ValueError(
+                    "Performing this reduction operation would produce "
+                    "a dense result: %s" % str(method)
+                )
+
+        if axis is None:
+            axis = tuple(range(self.ndim))
+
+        if not isinstance(axis, tuple):
+            axis = (axis,)
+
+        axis = tuple(a if a >= 0 else a + self.ndim for a in axis)
+        neg_axis = tuple(ax for ax in range(self.ndim) if ax not in set(axis))
+        a = self.transpose(neg_axis + axis)
+        a = a.reshape(
+            (
+                np.prod([self.shape[d] for d in neg_axis], dtype=np.intp),
+                np.prod([self.shape[d] for d in axis], dtype=np.intp),
+            )
+        )
+        data, inv_idx, counts = _grouped_reduce(a.data, a.coords[0], method, **kwargs)
+        result_fill_value = self.fill_value
+
+        if reduce_super_ufunc is None:
+            missing_counts = counts != a.shape[1]
+            data[missing_counts] = method(
+                data[missing_counts], self.fill_value, **kwargs
+            )
+        else:
+            data = method(
+                data, reduce_super_ufunc(self.fill_value, a.shape[1] - counts),
+            ).astype(data.dtype)
+            result_fill_value = reduce_super_ufunc(self.fill_value, a.shape[1])
+
+        coords = a.coords[0:1, inv_idx]
+        out = COO(
+            coords,
+            data,
+            shape=(a.shape[0],),
+            has_duplicates=False,
+            sorted=True,
+            prune=True,
+            fill_value=result_fill_value,
+        )
+        out = out.reshape(tuple(self.shape[d] for d in neg_axis))
+
+        if keepdims:
+            shape = list(self.shape)
+            for ax in axis:
+                shape[ax] = 1
+            out = out.reshape(shape)
+
+        if out.ndim == 0:
+            return out[()]
+
+        return out
 
     def mean(self, axis=None, keepdims=False, dtype=None, out=None):
         """
@@ -1833,3 +1960,61 @@ def as_coo(x, shape=None, fill_value=None):
         "Format not supported for conversion. Supplied type is "
         "%s, see help(sparse.as_coo) for supported formats." % type(x)
     )
+
+
+@numba.jit(nopython=True, nogil=True)  # pragma: no cover
+def _calc_counts_invidx(groups):
+    inv_idx = []
+    counts = []
+
+    if len(groups) == 0:
+        return (
+            np.array(inv_idx, dtype=groups.dtype),
+            np.array(counts, dtype=groups.dtype),
+        )
+
+    inv_idx.append(0)
+
+    last_group = groups[0]
+    for i in range(1, len(groups)):
+        if groups[i] != last_group:
+            counts.append(i - inv_idx[-1])
+            inv_idx.append(i)
+            last_group = groups[i]
+
+    counts.append(len(groups) - inv_idx[-1])
+
+    return (np.array(inv_idx, dtype=groups.dtype), np.array(counts, dtype=groups.dtype))
+
+
+def _grouped_reduce(x, groups, method, **kwargs):
+    """
+    Performs a :code:`ufunc` grouped reduce.
+
+    Parameters
+    ----------
+    x : np.ndarray
+        The data to reduce.
+    groups : np.ndarray
+        The groups the data belongs to. The groups must be
+        contiguous.
+    method : np.ufunc
+        The :code:`ufunc` to use to perform the reduction.
+    kwargs : dict
+        The kwargs to pass to the :code:`ufunc`'s :code:`reduceat`
+        function.
+
+    Returns
+    -------
+    result : np.ndarray
+        The result of the grouped reduce operation.
+    inv_idx : np.ndarray
+        The index of the first element where each group is found.
+    counts : np.ndarray
+        The number of elements in each group.
+    """
+    # Partial credit to @shoyer
+    # Ref: https://gist.github.com/shoyer/f538ac78ae904c936844
+    inv_idx, counts = _calc_counts_invidx(groups)
+    result = method.reduceat(x, inv_idx, **kwargs)
+    return result, inv_idx, counts
