+++ sparse/_sparse_array.py
@@ -334,162 +334,13 @@ class SparseArray:
         kwargs : dict
             Any extra arguments to pass to the reduction operation.
 
-        Returns
-        -------
-        GCXS
-            The result of the reduction operation.
-
-        Raises
-        ------
-        ValueError
-            If reducing an all-zero axis would produce a nonzero result.
-
         See Also
         --------
         numpy.ufunc.reduce : A similar Numpy method.
-        COO.reduce : Equivalent operation on COO arrays.
-
-        Examples
-        --------
-        You can use the :obj:`COO.reduce` method to apply a reduction operation to
-        any Numpy :code:`ufunc`.
-
-        >>> from sparse import COO
-        >>> x = np.ones((5, 5), dtype=np.int)
-        >>> s = COO.from_numpy(x)
-        >>> s2 = s.reduce(np.add, axis=1)
-        >>> s2.todense()  # doctest: +NORMALIZE_WHITESPACE
-        array([5, 5, 5, 5, 5])
-
-        You can also use the :code:`keepdims` argument to keep the dimensions after the
-        reduction.
-
-        >>> s3 = s.reduce(np.add, axis=1, keepdims=True)
-        >>> s3.shape
-        (5, 1)
-
-        You can also pass in any keyword argument that :obj:`numpy.ufunc.reduce` supports.
-        For example, :code:`dtype`. Note that :code:`out` isn't supported.
-
-        >>> s4 = s.reduce(np.add, axis=1, dtype=np.float16)
-        >>> s4.dtype
-        dtype('float16')
-
-        By default, this reduces the array by only the first axis.
-
-        >>> s.reduce(np.add)
-        <COO: shape=(5,), dtype=int64, nnz=5, fill_value=0>
+        COO.reduce : This method implemented on COO arrays.
+        GCXS.reduce : This method implemented on GCXS arrays.
         """
-        axis = normalize_axis(axis, self.ndim)
-        zero_reduce_result = method.reduce([self.fill_value, self.fill_value], **kwargs)
-        reduce_super_ufunc = None
-
-        if not equivalent(zero_reduce_result, self.fill_value):
-            reduce_super_ufunc = _reduce_super_ufunc.get(method, None)
-
-            if reduce_super_ufunc is None:
-                raise ValueError(
-                    "Performing this reduction operation would produce "
-                    "a dense result: %s" % str(method)
-                )
-
-        if axis is None:
-            if hasattr(self, "indptr"):  # gcxs
-                if reduce_super_ufunc is None:
-                    temp = method(self.data, self.fill_value)
-                    val = method.reduce(temp, **kwargs)
-                else:
-                    temp = method(
-                        self.data,
-                        reduce_super_ufunc(self.fill_value, self.size - self.nnz),
-                    )
-                    val = method.reduce(temp, **kwargs)
-                if keepdims:
-                    out = np.array(val)
-                    return out.reshape(np.ones(self.ndim, dtype=np.intp))
-                return val
-            else:  # coo
-                axis = tuple(range(self.ndim))
-
-        if not isinstance(axis, tuple):
-            axis = (axis,)
-
-        if hasattr(self, "indptr"):  # gcxs
-            r = np.arange(self.ndim, dtype=np.intp)
-            compressed_axes = [a for a in r if a not in set(axis)]
-            x = self.change_compressed_axes(compressed_axes)
-            idx = np.diff(x.indptr) != 0
-            indptr = x.indptr[:-1][idx]
-            indices = (np.arange(x._compressed_shape[0], dtype=np.intp))[idx]
-            data = method.reduceat(x.data, indptr, **kwargs)
-            counts = x.indptr[1:][idx] - x.indptr[:-1][idx]
-            n_cols = x._compressed_shape[1]
-
-        else:  # coo
-            axis = tuple(a if a >= 0 else a + self.ndim for a in axis)
-
-            neg_axis = tuple(ax for ax in range(self.ndim) if ax not in set(axis))
-
-            a = self.transpose(neg_axis + axis)
-            a = a.reshape(
-                (
-                    np.prod([self.shape[d] for d in neg_axis], dtype=np.intp),
-                    np.prod([self.shape[d] for d in axis], dtype=np.intp),
-                )
-            )
-            data, inv_idx, counts = _grouped_reduce(
-                a.data, a.coords[0], method, **kwargs
-            )
-            n_cols = a.shape[1]
-
-        result_fill_value = self.fill_value
-
-        if reduce_super_ufunc is None:
-            missing_counts = counts != n_cols
-            data[missing_counts] = method(
-                data[missing_counts], self.fill_value, **kwargs
-            )
-        else:
-            data = method(
-                data, reduce_super_ufunc(self.fill_value, n_cols - counts),
-            ).astype(data.dtype)
-            result_fill_value = reduce_super_ufunc(self.fill_value, n_cols)
-
-        if hasattr(self, "indptr"):  # gcxs
-            # prune data
-            mask = ~equivalent(data, result_fill_value)
-            data = data[mask]
-            indices = indices[mask]
-            out = type(self)(
-                (data, indices, []),
-                shape=(x._compressed_shape[0],),
-                fill_value=result_fill_value,
-                compressed_axes=None,
-            )
-            out = out.reshape(tuple(self.shape[d] for d in compressed_axes))
-        else:  # coo
-            coords = a.coords[0:1, inv_idx]
-            out = type(self)(
-                coords,
-                data,
-                shape=(a.shape[0],),
-                has_duplicates=False,
-                sorted=True,
-                prune=True,
-                fill_value=result_fill_value,
-            )
-            out = out.reshape(tuple(self.shape[d] for d in neg_axis))
-
-        if keepdims:
-            shape = list(self.shape)
-            for ax in axis:
-                shape[ax] = 1
-            out = out.reshape(shape)
-
-        if out.ndim == 0:
-            return out[()]
-
-        return out
+        raise NotImplementedError
 
     def sum(self, axis=None, keepdims=False, dtype=None, out=None):
         """
@@ -639,61 +490,3 @@ class SparseArray:
         return np.multiply.reduce(
             self, out=out, axis=axis, keepdims=keepdims, dtype=dtype
         )
-
-
-@numba.jit(nopython=True, nogil=True)  # pragma: no cover
-def _calc_counts_invidx(groups):
-    inv_idx = []
-    counts = []
-
-    if len(groups) == 0:
-        return (
-            np.array(inv_idx, dtype=groups.dtype),
-            np.array(counts, dtype=groups.dtype),
-        )
-
-    inv_idx.append(0)
-
-    last_group = groups[0]
-    for i in range(1, len(groups)):
-        if groups[i] != last_group:
-            counts.append(i - inv_idx[-1])
-            inv_idx.append(i)
-            last_group = groups[i]
-
-    counts.append(len(groups) - inv_idx[-1])
-
-    return (np.array(inv_idx, dtype=groups.dtype), np.array(counts, dtype=groups.dtype))
-
-
-def _grouped_reduce(x, groups, method, **kwargs):
-    """
-    Performs a :code:`ufunc` grouped reduce.
-
-    Parameters
-    ----------
-    x : np.ndarray
-        The data to reduce.
-    groups : np.ndarray
-        The groups the data belongs to. The groups must be
-        contiguous.
-    method : np.ufunc
-        The :code:`ufunc` to use to perform the reduction.
-    kwargs : dict
-        The kwargs to pass to the :code:`ufunc`'s :code:`reduceat`
-        function.
-
-    Returns
-    -------
-    result : np.ndarray
-        The result of the grouped reduce operation.
-    inv_idx : np.ndarray
-        The index of the first element where each group is found.
-    counts : np.ndarray
-        The number of elements in each group.
-    """
-    # Partial credit to @shoyer
-    # Ref: https://gist.github.com/shoyer/f538ac78ae904c936844
-    inv_idx, counts = _calc_counts_invidx(groups)
-    result = method.reduceat(x, inv_idx, **kwargs)
-    return result, inv_idx, counts
